# [CWS] preallocate slice in `UnmarshalStringArray`