<!--
* Contributors are encouraged to read our [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
* Please fill the below sections if possible with relevant information or links.
-->
### What does this PR do?

When unmarshalling a string array the allocation cost is split into 2:
- the allocation of the underlying string values
- the allocation of the slice itself

This PR optimizes this second point by doing a first pass on the byte slice to pre-compute the final size before reading the string themselves.

Benchmark:
```
                        │   old.txt   │              new.txt               │
                        │   sec/op    │   sec/op     vs base               │
UnmarshalStringArray-10   81.69µ ± 5%   74.71µ ± 3%  -8.55% (p=0.000 n=10)

                        │    old.txt    │               new.txt                │
                        │     B/op      │     B/op      vs base                │
UnmarshalStringArray-10   250.37Ki ± 0%   80.01Ki ± 0%  -68.04% (p=0.000 n=10)

                        │   old.txt   │              new.txt               │
                        │  allocs/op  │  allocs/op   vs base               │
UnmarshalStringArray-10   4.111k ± 0%   4.097k ± 0%  -0.34% (p=0.000 n=10)
```

### Motivation

### Describe how you validated your changes
<!--
Validate your changes before merge, ensuring that:
* Your PR is tested by static / unit / integrations / e2e tests
* Your PR description details which e2e tests cover your changes, if any
* The PR description contains details of how you validated your changes. If you validated changes manually and not through automated tests, add context on why automated tests did not fit your changes validation.

If you want additional validation by a second person, you can ask reviewers to do it. Describe how to set up an environment for manual tests in the PR description. Manual validation is expected to happen on every commit before merge.

Any manual validation step should then map to an automated test. Manual validation should not substitute automation, minus exceptions not supported by test tooling yet.
-->

### Possible Drawbacks / Trade-offs

### Additional Notes
<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->